MAT-7016: Move Special Character error message const to be colated which validation logic. Having it located in the QDM components and used by the QI-Core components causes QI-Core clause highlighting to lose styling, for some reason.
